+++ itest/babylon/babylon_e2e_test.go
@@ -234,7 +234,9 @@ func TestCatchingUp(t *testing.T) {
 	t.Logf("the latest finalized block is at %v", finalizedBlock.GetHeight())
 
 	// check if the fast sync works by checking if the gap is not more than 1
-	currentHeight, err := tm.BBNConsumerClient.QueryLatestBlockHeight(ctx)
+	currentBlock, err := tm.BBNConsumerClient.QueryLatestBlock(ctx)
+	require.NoError(t, err)
+	currentHeight := currentBlock.GetHeight()
 	t.Logf("the current block is at %v", currentHeight)
 	require.NoError(t, err)
 	require.True(t, currentHeight < finalizedBlock.GetHeight()+uint64(n))