+++ CHANGELOG.md
@@ -56,6 +56,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [537](https://github.com/babylonlabs-io/finality-provider/pull/537) chore: mv helper methods from fpi
 * [#538](https://github.com/babylonlabs-io/finality-provider/pull/538) chore: fix e2e tags and wrap check
 * [520](https://github.com/babylonlabs-io/finality-provider/pull/520) chore: migration guide keyring test to file
+* [#539](https://github.com/babylonlabs-io/finality-provider/pull/539) api: replace QueryLatestBlockHeight with QueryLatestBlock
+* [#547](https://github.com/babylonlabs-io/finality-provider/pull/547) api: consolidate QueryActivatedHeight and QueryFinalityActivatationHeight APIs
 * [542](https://github.com/babylonlabs-io/finality-provider/pull/542) chore: abstract finality submitter
 
 ## v1.1.0-rc.1